requirements.txt
# Core runtime
streamlit>=1.36.0
openai>=1.40.0
chromadb>=0.5.4
tiktoken>=0.7.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.7.0
tenacity>=8.3.0
